+++ jsonschema-generator/src/main/java/com/github/victools/jsonschema/generator/SchemaGeneratorConfigBuilder.java
@@ -229,7 +229,7 @@ public class SchemaGeneratorConfigBuilder {
     }
 
     /**
-     * Adding a custom schema provider – if it returns null for a given type, the next definition provider will be applied.
+     * Adding a custom schema provider - if it returns null for a given type, the next definition provider will be applied.
      * <br>
      * If all custom schema providers return null (or there is none), then the standard behaviour applies.
      *
@@ -244,9 +244,9 @@ public class SchemaGeneratorConfigBuilder {
     }
 
     /**
-     * Adding an override for type attributes – all of the registered overrides will be applied in the order of having been added.
+     * Adding an override for type attributes - all the registered overrides will be applied in the order of having been added.
      *
-     * @param override adding/removing attributes on a JSON Schema node – specifically intended for attributes relating to the type in general.
+     * @param override adding/removing attributes on a JSON Schema node - specifically intended for attributes relating to the type in general.
      * @return this builder instance (for chaining)
      * @deprecated use {@link SchemaGeneratorGeneralConfigPart#withTypeAttributeOverride(TypeAttributeOverrideV2)} instead
      */