+++ jsonschema-generator/src/main/java/com/github/victools/jsonschema/generator/impl/module/SimpleTypeModule.java
@@ -232,7 +232,7 @@ public class SimpleTypeModule implements Module {
     }
 
     /**
-     * Determine whether a given type is nullable – returning false if the type refers to a primitive type.
+     * Determine whether a given type is nullable - returning false if the type refers to a primitive type.
      *
      * @param fieldOrMethod reference to the method/field (which is being ignored here)
      * @return false if type is a primitive, otherwise null
@@ -283,13 +283,14 @@ public class SimpleTypeModule implements Module {
     }
 
     private boolean shouldHaveEmptySchema(TypeScope scope) {
-        return SchemaKeyword.TAG_TYPE_NULL == this.fixedJsonSchemaTypes.get(scope.getType().getErasedType());
+        return scope.getType() == null
+               || SchemaKeyword.TAG_TYPE_NULL == this.fixedJsonSchemaTypes.get(scope.getType().getErasedType());
     }
 
     /**
      * Implementation of the {@link CustomDefinitionProviderV2} interface for applying fixed schema definitions for simple java types.
      */
-    private class SimpleTypeDefinitionProvider implements CustomDefinitionProviderV2 {
+    private final class SimpleTypeDefinitionProvider implements CustomDefinitionProviderV2 {
 
         @Override
         public CustomDefinition provideCustomSchemaDefinition(ResolvedType javaType, SchemaGenerationContext context) {